Move type variables to a single file.
