+++ python/pyspark/pandas/extensions.py
@@ -14,16 +14,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from typing import Callable, Generic, Optional, Type, TypeVar, Union, TYPE_CHECKING
+from typing import Callable, Generic, Optional, Type, Union, TYPE_CHECKING
 import warnings
+
+from pyspark.pandas._typing import T
 
 if TYPE_CHECKING:
     from pyspark.pandas.frame import DataFrame  # noqa: F401 (SPARK-34943)
     from pyspark.pandas.indexes import Index  # noqa: F401 (SPARK-34943)
     from pyspark.pandas.series import Series  # noqa: F401 (SPARK-34943)
-
-
-T = TypeVar("T")
 
 
 class CachedAccessor(Generic[T]):
